refactor: maintained problematic code
