--- src/components/filter-results/create-element/create-forme-element.js
@@ -1,9 +1,0 @@
-export function judgeFormeContent(data, pokemonFormResult) {
-  if (data.baseForme) {
-    pokemonFormResult.textContent = data.baseForme;
-  } else if (data.forme) {
-    pokemonFormResult.textContent = data.forme;
-  } else {
-    pokemonFormResult.textContent = '';
-  }
-}
