+++ src/components/filter-results/filter-results.scss
@@ -437,6 +437,12 @@
         width: 64px;
         overflow: hidden;
       }
+
+      .stat-bst {
+        p {
+          font-weight: 800;
+        }
+      }
     }
 
     .pokemon-moves {
