+++ src/constants/pokemon-categories-list.js
@@ -1,0 +1,339 @@
+const UNRESTRICTED_LEGENGARY = [
+  'articuno',
+  'articunogalar',
+  'zapdos',
+  'zapdosgalar',
+  'moltres',
+  'moltresgalar',
+  'raikou',
+  'entei',
+  'suicune',
+  'regirock',
+  'regiice',
+  'registeel',
+  'latias',
+  'latios',
+  'uxie',
+  'mesprit',
+  'azelf',
+  'heatran',
+  'regigigas',
+  'cresselia',
+  'cobalion',
+  'terrakion',
+  'virizion',
+  'tornadus',
+  'tornadustherian',
+  'thundurus',
+  'thundurustherian',
+  'landorus',
+  'landorustherian',
+  'tapukoko',
+  'tapulele',
+  'tapubulu',
+  'tapufini',
+  'kubfu',
+  'urshifu',
+  'urshifurapidstrike',
+  'regieleki',
+  'regidrago',
+  'glastrier',
+  'spectrier',
+  'enamorus',
+  'enamorustherian',
+  'wochien',
+  'chienpao',
+  'tinglu',
+  'chiyu',
+  'okidogi',
+  'munkidori',
+  'fezandipiti',
+  'ogerpon',
+  'ogerponwellspring',
+  'ogerponhearthflame',
+  'ogerponcornerstone'
+];
+
+const RESTRICTED_LEGENGARY = [
+  'mewtwo',
+  'lugia',
+  'hooh',
+  'kyogre',
+  'groudon',
+  'rayquaza',
+  'dialga',
+  'dialgaorigin',
+  'palkia',
+  'palkiaorigin',
+  'giratina',
+  'giratinaorigin',
+  'reshiram',
+  'zekrom',
+  'kyurem',
+  'kyuremblack',
+  'kyuremwhite',
+  'xerneas',
+  'yveltal',
+  'zygarde',
+  'zygarde10',
+  'zygardecomplete',
+  'zygardecomplete',
+  'cosmog',
+  'cosmoem',
+  'solgaleo',
+  'lunala',
+  'necrozma',
+  'necrozmaduskmane',
+  'necrozmadawnwings',
+  'zacian',
+  'zaciancrowned',
+  'zamazenta',
+  'zamazentacrowned',
+  'eternatus',
+  'calyrex',
+  'calyrexice',
+  'calyrexshadow',
+  'koraidon',
+  'miraidon',
+  'terapagos',
+  'terapagos',
+  'terapagosterastal',
+  'terapagosstellar'
+];
+
+const MYTHICAL = [
+  'mew',
+  'celebi',
+  'celebi',
+  'jirachi',
+  'jirachi',
+  'deoxys',
+  'deoxysattack',
+  'deoxysdefense',
+  'deoxysspeed',
+  'phione',
+  'manaphy',
+  'darkrai',
+  'shaymin',
+  'shayminsky',
+  'shayminsky',
+  'arceus',
+  'victini',
+  'keldeo',
+  'keldeoresolute',
+  'meloetta',
+  'meloettapirouette',
+  'meloettapirouette',
+  'genesect',
+  'diancie',
+  'diancie',
+  'hoopa',
+  'hoopaunbound',
+  'volcanion',
+  'volcanion',
+  'magearna',
+  'marshadow',
+  'zeraora',
+  'meltan',
+  'melmetal',
+  'melmetal',
+  'zarude',
+  'pecharunt'
+];
+
+const LATE_BLOOMER = [
+  'dragonite',
+  'tyranitar',
+  'salamence',
+  'metagross',
+  'garchomp',
+  'hydreigon',
+  'goodra',
+  'goodrahisui',
+  'goodrahisui',
+  'kommo',
+  'dragapult',
+  'baxcalibur'
+];
+
+const ULTRA_BEAST = [
+  'nihilego',
+  'buzzwole',
+  'pheromosa',
+  'xurkitree',
+  'celesteela',
+  'kartana',
+  'guzzlord',
+  'poipole',
+  'naganadel',
+  'stakataka',
+  'blacephalon'
+];
+
+const PARADOX = [
+  'greattusk',
+  'screamtail',
+  'brutebonnet',
+  'fluttermane',
+  'slitherwing',
+  'sandyshocks',
+  'roaringmoon',
+  'Walkingwake',
+  'GougingFire',
+  'RagingBolt',
+  'IronTreads',
+  'IronBundle',
+  'IronHands',
+  'IronJugulis',
+  'IronMoth',
+  'IronThorns',
+  'IronValiant',
+  'IronLeaves',
+  'IronBoulder',
+  'IronCrown'
+];
+
+const INVALID_FORMS = [
+  'arceusbug',
+  'arceusdark',
+  'arceusdragon',
+  'arceuselectric',
+  'arceusfairy',
+  'arceusfighting',
+  'arceusfire',
+  'arceusflying',
+  'arceusghost',
+  'arceusgrass',
+  'arceusground',
+  'arceusice',
+  'arceuspoison',
+  'arceuspsychic',
+  'arceusrock',
+  'arceussteel',
+  'arceuswater',
+  'cherrimsunshine',
+  'cramorantgorging',
+  'cramorantgulping',
+  'dudunsparcethreesegment',
+  'eeveestarter',
+  'eternamax',
+  'floetteeternal',
+  'gmax',
+  'genesectdouse',
+  'genesectshock',
+  'genesectburn',
+  'genesectchill',
+  'greninjaash',
+  'keldeoresolute',
+  'magearnaoriginal',
+  'mausholdfour',
+  'mega',
+  'mimikyubusted',
+  'missingno',
+  'morpekohangry',
+  'necrozmaultra',
+  'ogerpontealtera',
+  'ogerponwellspringtera',
+  'ogerponhearthflametera',
+  'ogerponcornerstonetera',
+  'pichuspikyeared',
+  'pikachucosplay',
+  'pikachurockstar',
+  'pikachubelle',
+  'pikachupopstar',
+  'pikachuphd',
+  'pikachulibre',
+  'pikachuoriginal',
+  'pikachuhoenn',
+  'pikachusinnoh',
+  'pikachuunova',
+  'pikachukalos',
+  'pikachualola',
+  'pikachupartner',
+  'pikachustarter',
+  'pikachuworld',
+  'poltchageistartisan',
+  'polteageistantique',
+  'primal',
+  'silvallybug',
+  'silvallydark',
+  'silvallydragon',
+  'silvallyelectric',
+  'silvallyfairy',
+  'silvallyfighting',
+  'silvallyfire',
+  'silvallyflying',
+  'silvallyghost',
+  'silvallygrass',
+  'silvallyground',
+  'silvallyice',
+  'silvallypoison',
+  'silvallypsychic',
+  'silvallyrock',
+  'silvallysteel',
+  'silvallywater',
+  'sinistchamasterpiece',
+  'sinisteaantique',
+  'totem',
+  'vivillonfancy',
+  'vivillonpokeball',
+  'xerneasneutral',
+  'zarudedada'
+];
+
+const UNUSUAL_FORME_BAN_LIST = [
+  'aegislashblade',
+  'castformrainy',
+  'castformsnowy',
+  'castformsunny',
+  'darmanitanzen',
+  'darmanitangalarzen',
+  'gourgeistlarge',
+  'gourgeistsmall',
+  'gourgeistsuper',
+  'pumpkaboolarge',
+  'pumpkaboosmall',
+  'pumpkaboosuper',
+  'wishiwashischool'
+];
+
+const EXCEPTIONS = ['yanmega', 'meganium'];
+
+export const specialPokemons = [
+  {
+    id: 'invalidForms',
+    data: INVALID_FORMS
+  },
+  {
+    id: 'exceptions',
+    data: EXCEPTIONS
+  },
+  {
+    id: 'unusualFormsBanList',
+    data: UNUSUAL_FORME_BAN_LIST
+  },
+  {
+    id: 'unrestrictedLegengary',
+    data: UNRESTRICTED_LEGENGARY
+  },
+  {
+    id: 'restrictedLegengary',
+    data: RESTRICTED_LEGENGARY
+  },
+  {
+    id: 'mythical',
+    data: MYTHICAL
+  },
+  {
+    id: 'lateBloomer',
+    data: LATE_BLOOMER
+  },
+  {
+    id: 'ultraBeast',
+    data: ULTRA_BEAST
+  },
+  {
+    id: 'paradox',
+    data: PARADOX
+  }
+];
