+++ src/filtered-pokemon-list.js
@@ -1,305 +1,240 @@
 const UNRESTRICTED_LEGENGARY = [
-  'articuno',
-  'articunogalar',
-  'zapdos',
-  'zapdosgalar',
-  'moltres',
-  'moltresgalar',
-  'raikou',
-  'entei',
-  'suicune',
-  'regirock',
-  'regiice',
-  'registeel',
-  'latias',
-  'latios',
-  'uxie',
-  'mesprit',
-  'azelf',
-  'heatran',
-  'regigigas',
-  'cresselia',
-  'cobalion',
-  'terrakion',
-  'virizion',
-  'tornadus',
-  'tornadustherian',
-  'thundurus',
-  'thundurustherian',
-  'landorus',
-  'landorustherian',
-  'tapukoko',
-  'tapulele',
-  'tapubulu',
-  'tapufini',
-  'kubfu',
-  'urshifu',
-  'urshifurapidstrike',
-  'regieleki',
-  'regidrago',
-  'glastrier',
-  'spectrier',
-  'enamorus',
-  'enamorustherian',
-  'wochien',
-  'chienpao',
-  'tinglu',
-  'chiyu',
-  'okidogi',
-  'munkidori',
-  'fezandipiti',
-  'ogerpon',
-  'ogerponwellspring',
-  'ogerponhearthflame',
-  'ogerponcornerstone'
+  'Articuno',
+  'Articuno-Galar',
+  'Zapdos',
+  'Zapdos-Galar',
+  'Moltres',
+  'Moltres-Galar',
+  'Raikou',
+  'Entei',
+  'Suicune',
+  'Regirock',
+  'Regiice',
+  'Registeel',
+  'Latias',
+  'Latios',
+  'Uxie',
+  'Mesprit',
+  'Azelf',
+  'Heatran',
+  'Regigigas',
+  'Cresselia',
+  'Cobalion',
+  'Terrakion',
+  'Virizion',
+  'Tornadus',
+  'Tornadus-Therian',
+  'Thundurus',
+  'Thundurus-Therian',
+  'Landorus',
+  'Landorus-Therian',
+  'Tapu Koko',
+  'Tapu Lele',
+  'Tapu Bulu',
+  'Tapu Fini',
+  'Kubfu',
+  'Urshifu',
+  'Urshifu-Rapid-Strike',
+  'Regieleki',
+  'Regidrago',
+  'Glastrier',
+  'Spectrier',
+  'Enamorus',
+  'Enamorus-Therian',
+  'Wo-Chien',
+  'Chien-Pao',
+  'Ting-Lu',
+  'Chi-Yu',
+  'Okidogi',
+  'Munkidori',
+  'Fezandipiti',
+  'Ogerpon',
+  'Ogerpon-Wellspring',
+  'Ogerpon-Hearthflame',
+  'Ogerpon-Cornerstone'
 ];
 
 const RESTRICTED_LEGENGARY = [
-  'mewtwo',
-  'lugia',
-  'hooh',
-  'kyogre',
-  'groudon',
-  'rayquaza',
-  'dialga',
-  'dialgaorigin',
-  'palkia',
-  'palkiaorigin',
-  'giratina',
-  'giratinaorigin',
-  'reshiram',
-  'zekrom',
-  'kyurem',
-  'kyuremblack',
-  'kyuremwhite',
-  'xerneas',
-  'yveltal',
-  'zygarde',
-  'zygarde10',
-  'zygardecomplete',
-  'zygardecomplete',
-  'cosmog',
-  'cosmoem',
-  'solgaleo',
-  'lunala',
-  'necrozma',
-  'necrozmaduskmane',
-  'necrozmadawnwings',
-  'zacian',
-  'zaciancrowned',
-  'zamazenta',
-  'zamazentacrowned',
-  'eternatus',
-  'calyrex',
-  'calyrexice',
-  'calyrexshadow',
-  'koraidon',
-  'miraidon',
-  'terapagos',
-  'terapagos',
-  'terapagosterastal',
-  'terapagosstellar'
+  'Mewtwo',
+  'Lugia',
+  'Ho-Oh',
+  'Kyogre',
+  'Groudon',
+  'Rayquaza',
+  'Dialga',
+  'Dialga-Origin',
+  'Palkia',
+  'Palkiao-Origin',
+  'Giratina',
+  'Giratina-Origin',
+  'Reshiram',
+  'Zekrom',
+  'Kyurem',
+  'Kyurem-Black',
+  'Kyurem-White',
+  'Xerneas',
+  'Yveltal',
+  'Zygarde',
+  'Zygarde-10',
+  'Zygarde-Complete',
+  'Cosmog',
+  'Cosmoem',
+  'Solgaleo',
+  'Lunala',
+  'Necrozma',
+  'Necrozma-Dusk-Mane',
+  'Necrozma-Dawn-Wings',
+  'Zacian',
+  'Zacian-Crowned',
+  'Zamazenta',
+  'Zamazenta-Crowned',
+  'Eternatus',
+  'Calyrex',
+  'Calyrex-Ice',
+  'Calyrex-Shadow',
+  'Koraidon',
+  'Miraidon',
+  'Terapagos',
+  'Terapagos-Terastal',
+  'Terapagos-Stellar'
 ];
 
 const MYTHICAL = [
-  'mew',
-  'celebi',
-  'celebi',
-  'jirachi',
-  'jirachi',
-  'deoxys',
-  'deoxysattack',
-  'deoxysdefense',
-  'deoxysspeed',
-  'phione',
-  'manaphy',
-  'darkrai',
-  'shaymin',
-  'shayminsky',
-  'shayminsky',
-  'arceus',
-  'victini',
-  'keldeo',
-  'keldeoresolute',
-  'meloetta',
-  'meloettapirouette',
-  'meloettapirouette',
-  'genesect',
-  'diancie',
-  'diancie',
-  'hoopa',
-  'hoopaunbound',
-  'volcanion',
-  'volcanion',
-  'magearna',
-  'marshadow',
-  'zeraora',
-  'meltan',
-  'melmetal',
-  'melmetal',
-  'zarude',
-  'pecharunt'
+  'Mew',
+  'Celebi',
+  'Jirachi',
+  'Deoxys',
+  'Deoxys-',
+  'Phione',
+  'Manaphy',
+  'Darkrai',
+  'Shaymin',
+  'Shaymin-Sky',
+  'Arceus',
+  'Victini',
+  'Keldeo',
+  'Keldeoresolute',
+  'Meloetta',
+  'Meloetta-Pirouette',
+  'Genesect',
+  'Diancie',
+  'Hoopa',
+  'Hoopa-Unbound',
+  'Volcanion',
+  'Magearna',
+  'Marshadow',
+  'Zeraora',
+  'Meltan',
+  'Melmetal',
+  'Zarude',
+  'Pecharunt'
 ];
 
 const LATE_BLOOMER = [
-  'dragonite',
-  'tyranitar',
-  'salamence',
-  'metagross',
-  'garchomp',
-  'hydreigon',
-  'goodra',
-  'goodrahisui',
-  'goodrahisui',
-  'kommo',
-  'dragapult',
-  'baxcalibur'
+  'Dragonite',
+  'Tyranitar',
+  'Salamence',
+  'Metagross',
+  'Garchomp',
+  'Hydreigon',
+  'Goodra',
+  'Goodra-Hisui',
+  'Kommo-o',
+  'Dragapult',
+  'Baxcalibur'
 ];
 
 const ULTRA_BEAST = [
-  'nihilego',
-  'buzzwole',
-  'pheromosa',
-  'xurkitree',
-  'celesteela',
-  'kartana',
-  'guzzlord',
-  'poipole',
-  'naganadel',
-  'stakataka',
-  'blacephalon'
+  'Nihilego',
+  'Buzzwole',
+  'Pheromosa',
+  'Xurkitree',
+  'Celesteela',
+  'Kartana',
+  'Guzzlord',
+  'Poipole',
+  'Naganadel',
+  'Stakataka',
+  'Blacephalon'
 ];
 
 const PARADOX = [
-  'greattusk',
-  'screamtail',
-  'brutebonnet',
-  'fluttermane',
-  'slitherwing',
-  'sandyshocks',
-  'roaringmoon',
-  'Walkingwake',
-  'GougingFire',
-  'RagingBolt',
-  'IronTreads',
-  'IronBundle',
-  'IronHands',
-  'IronJugulis',
-  'IronMoth',
-  'IronThorns',
-  'IronValiant',
-  'IronLeaves',
-  'IronBoulder',
-  'IronCrown'
+  'Great Tusk',
+  'Scream Tail',
+  'Brute Bonnet',
+  'Flutter Mane',
+  'Slither Wing',
+  'Sandy Shocks',
+  'Roaring Moon',
+  'Walking Wake',
+  'Gouging Fire',
+  'Raging Bolt',
+  'Iron Treads',
+  'Iron Bundle',
+  'Iron Hands',
+  'Iron Jugulis',
+  'Iron Moth',
+  'Iron Thorns',
+  'Iron Valiant',
+  'Iron Leaves',
+  'Iron Boulder',
+  'Iron Crown'
 ];
 
 const INVALID_FORMS = [
-  'arceusbug',
-  'arceusdark',
-  'arceusdragon',
-  'arceuselectric',
-  'arceusfairy',
-  'arceusfighting',
-  'arceusfire',
-  'arceusflying',
-  'arceusghost',
-  'arceusgrass',
-  'arceusground',
-  'arceusice',
-  'arceuspoison',
-  'arceuspsychic',
-  'arceusrock',
-  'arceussteel',
-  'arceuswater',
-  'cherrimsunshine',
-  'cramorantgorging',
-  'cramorantgulping',
-  'dudunsparcethreesegment',
-  'eeveestarter',
-  'eternamax',
-  'floetteeternal',
-  'gmax',
-  'genesectdouse',
-  'genesectshock',
-  'genesectburn',
-  'genesectchill',
-  'greninjaash',
-  'keldeoresolute',
-  'magearnaoriginal',
-  'mausholdfour',
+  'Arceus-',
+  'Cherrim-Sunshine',
+  'Cramorant-Gorging',
+  'Cramorant-Gulping',
+  'Eevee-Starter',
+  'Genesect-',
+  'Greninja-Ash',
+  'Keldeo-Resolute',
+  'Magearna-Original',
+  'Maushold-Four',
   'mega',
-  'mimikyubusted',
-  'missingno',
+  'Mimikyu-Busted',
+  'Missingno',
   'morpekohangry',
-  'necrozmaultra',
-  'ogerpontealtera',
-  'ogerponwellspringtera',
-  'ogerponhearthflametera',
-  'ogerponcornerstonetera',
-  'pichuspikyeared',
-  'pikachucosplay',
-  'pikachurockstar',
-  'pikachubelle',
-  'pikachupopstar',
-  'pikachuphd',
-  'pikachulibre',
-  'pikachuoriginal',
-  'pikachuhoenn',
-  'pikachusinnoh',
-  'pikachuunova',
-  'pikachukalos',
-  'pikachualola',
-  'pikachupartner',
-  'pikachustarter',
-  'pikachuworld',
-  'poltchageistartisan',
-  'polteageistantique',
-  'primal',
-  'silvallybug',
-  'silvallydark',
-  'silvallydragon',
-  'silvallyelectric',
-  'silvallyfairy',
-  'silvallyfighting',
-  'silvallyfire',
-  'silvallyflying',
-  'silvallyghost',
-  'silvallygrass',
-  'silvallyground',
-  'silvallyice',
-  'silvallypoison',
-  'silvallypsychic',
-  'silvallyrock',
-  'silvallysteel',
-  'silvallywater',
-  'sinistchamasterpiece',
-  'sinisteaantique',
+  'Necrozma-Ultra',
+  'Ogerpon-Teal-Tera',
+  'Ogerpon-Wellspring-Tera',
+  'Ogerpon-Hearthflame-Tera',
+  'Ogerpon-Cornerstone-Tera',
+  'Pichu-Spikyeared',
+  'Pikachu-',
+  'Poltchageist-',
+  'Polteageist-',
+  'Silvally-',
+  'Sinistcha-',
+  'Sinistea-',
   'totem',
-  'vivillonfancy',
-  'vivillonpokeball',
-  'xerneasneutral',
-  'zarudedada'
+  'Vivillon-',
+  'Xerneas-',
+  'Zarude-'
 ];
 
-const unusualFormeBanList = [
-  'aegislashblade',
-  'castformrainy',
-  'castformsnowy',
-  'castformsunny',
-  'darmanitanzen',
-  'darmanitangalarzen',
-  'gourgeistlarge',
-  'gourgeistsmall',
-  'gourgeistsuper',
-  'pumpkaboolarge',
-  'pumpkaboosmall',
-  'pumpkaboosuper',
-  'wishiwashischool'
+const UNUSUAL_FORME_BAN_LIST = [
+  'Aegislash-Blade',
+  'Castform-Rainy',
+  'Castform-Snowy',
+  'Castform-Sunny',
+  'Darmanitan-Zen',
+  'Darmanitan-Galar-Zen',
+  'Dudunsparce-Three-Segment',
+  'Gourgeist-Large',
+  'Gourgeist-Small',
+  'Gourgeist-Super',
+  'Morpeko-Hangry',
+  'Pumpkaboo-Large',
+  'Pumpkaboo-Small',
+  'Pumpkaboo-Super',
+  'Wishiwashi-School'
 ];
 
-const EXCEPTIONS = ['yanmega', 'meganium'];
+const EXCEPTIONS = ['Yanmega'];
 
-const SpecialPokemons = {
+const specialPokemons = {
   unrestrictedLegengary: {
     name: 'unrestrictedLegengary',
     data: UNRESTRICTED_LEGENGARY
@@ -326,7 +261,7 @@ const SpecialPokemons = {
   }
 };
 
-const specialPokemonCategories = [SpecialPokemons['mythical'], SpecialPokemons['restrictedLegengary']];
+const specialPokemonCategories = [specialPokemons.mythical, specialPokemons.restrictedLegengary];
 
 function filterCondition(pokemonName) {
   return (
@@ -341,7 +276,7 @@ export function isValidPokemon(pokemonName, serialNumber) {
   return (
     serialNumber > 0 &&
     !filteredInvalidForms.some((form) => pokemonName.includes(form)) &&
-    !unusualFormeBanList.some((form) => pokemonName.includes(form)) &&
+    !UNUSUAL_FORME_BAN_LIST.some((form) => pokemonName.includes(form)) &&
     !filterCondition(pokemonName)
   );
 }
